+++ crates/core/action-pipeline/src/subscribers/moonbase.rs
@@ -164,12 +164,7 @@ impl Subscriber for MoonbaseSubscriber {
                     let touched_files = context
                         .touched_files
                         .iter()
-                        .map(|f| {
-                            f.strip_prefix(&workspace.root)
-                                .unwrap_or(f)
-                                .to_string_lossy()
-                                .to_string()
-                        })
+                        .map(|f| f.to_string_lossy().to_string())
                         .collect::<Vec<_>>();
 
                     let response = match graphql::post_mutation::<create_run::ResponseData>(