+++ crates/core/project-graph/src/project_builder.rs
@@ -597,8 +597,7 @@ impl<'ws> ProjectGraphBuilder<'ws> {
 
         for glob in globs {
             normalized_globs.push(glob::normalize(
-                //  glob.strip_prefix(&self.workspace.root).unwrap(),
-                glob,
+                glob.strip_prefix(&self.workspace.root).unwrap(),
             )?);
         }