+++ crates/core/task/src/task.rs
@@ -44,8 +44,10 @@ pub struct Task {
 
     pub inputs: Vec<InputValue>,
 
+    // Relative from workspace root
     pub input_globs: FxHashSet<FileGlob>,
 
+    // Absolute paths
     pub input_paths: FxHashSet<PathBuf>,
 
     pub input_vars: FxHashSet<String>,
@@ -57,8 +59,10 @@ pub struct Task {
 
     pub outputs: Vec<FilePath>,
 
+    // Relative from workspace root
     pub output_globs: FxHashSet<FileGlob>,
 
+    // Absolute paths
     pub output_paths: FxHashSet<PathBuf>,
 
     pub platform: PlatformType,
@@ -173,18 +177,7 @@ impl Task {
 
     /// Create a globset of all input globs to match with.
     pub fn create_globset(&self) -> Result<glob::GlobSet, TaskError> {
-        Ok(glob::GlobSet::new(
-            self.input_globs
-                .iter()
-                .map(|g| {
-                    if cfg!(windows) {
-                        glob::remove_drive_prefix(g)
-                    } else {
-                        g.to_owned()
-                    }
-                })
-                .collect::<Vec<String>>(),
-        )?)
+        Ok(glob::GlobSet::new(&self.input_globs)?)
     }
 
     /// Determine the type of task after inheritance and expansion.
@@ -202,7 +195,8 @@ impl Task {
     pub fn get_affected_files(
         &self,
         touched_files: &TouchedFilePaths,
-        project_root: &Path,
+        workspace_root: &Path,
+        project_source: &str,
     ) -> Result<Vec<PathBuf>, TaskError> {
         let mut files = vec![];
         let has_globs = !self.input_globs.is_empty();
@@ -210,13 +204,15 @@ impl Task {
 
         for file in touched_files {
             // Don't run on files outside of the project
-            if !file.starts_with(project_root) {
+            if !file.starts_with(project_source) {
                 continue;
             }
 
-            if self.input_paths.contains(file) || (has_globs && globset.matches(file)?) {
+            let abs_file = workspace_root.join(file);
+
+            if self.input_paths.contains(&abs_file) || (has_globs && globset.matches(file)?) {
                 // Mimic relative from ("./")
-                files.push(PathBuf::from(".").join(file.strip_prefix(project_root).unwrap()));
+                files.push(PathBuf::from(".").join(file.strip_prefix(project_source).unwrap()));
             }
         }
 
@@ -225,7 +221,11 @@ impl Task {
 
     /// Return true if this task is affected based on touched files.
     /// Will attempt to find any file that matches our list of inputs.
-    pub fn is_affected(&self, touched_files: &TouchedFilePaths) -> Result<bool, TaskError> {
+    pub fn is_affected(
+        &self,
+        touched_files: &TouchedFilePaths,
+        workspace_root: &Path,
+    ) -> Result<bool, TaskError> {
         for var_name in &self.input_vars {
             if let Ok(var) = env::var(var_name) {
                 if !var.is_empty() {
@@ -244,11 +244,13 @@ impl Task {
         let globset = self.create_globset()?;
 
         for file in touched_files {
-            if self.input_paths.contains(file) {
+            let abs_file = workspace_root.join(file);
+
+            if self.input_paths.contains(&abs_file) {
                 trace!(
                     target: self.get_log_target(),
                     "Affected by {} (via input files)",
-                    color::path(file),
+                    color::path(&abs_file),
                 );
 
                 return Ok(true);
@@ -258,7 +260,7 @@ impl Task {
                 trace!(
                     target: self.get_log_target(),
                     "Affected by {} (via input globs)",
-                    color::path(file),
+                    color::path(&abs_file),
                 );
 
                 return Ok(true);