+++ crates/core/task/tests/task_test.rs
@@ -2,9 +2,10 @@ use moon_config::{TaskCommandArgs, TaskConfig, TaskOptionEnvFileConfig, TaskOpti
 use moon_target::Target;
 use moon_task::{Task, TaskOptions};
 use moon_test_utils::{create_sandbox, get_fixtures_path};
-use moon_utils::{glob, string_vec};
+use moon_utils::string_vec;
 use rustc_hash::FxHashSet;
 use std::env;
+use std::path::PathBuf;
 
 pub fn create_task(config: TaskConfig) -> Task {
     Task::from_config(Target::new("project", "task").unwrap(), &config).unwrap()
@@ -339,7 +340,9 @@ mod is_affected {
 
         env::set_var("FOO", "foo");
 
-        assert!(task.is_affected(&FxHashSet::default()).unwrap());
+        assert!(task
+            .is_affected(&FxHashSet::default(), &PathBuf::new())
+            .unwrap());
 
         env::remove_var("FOO");
     }
@@ -353,7 +356,9 @@ mod is_affected {
 
         task.input_vars.insert("BAR".into());
 
-        assert!(!task.is_affected(&FxHashSet::default()).unwrap());
+        assert!(!task
+            .is_affected(&FxHashSet::default(), &PathBuf::new())
+            .unwrap());
     }
 
     #[test]
@@ -367,7 +372,9 @@ mod is_affected {
 
         env::set_var("BAZ", "");
 
-        assert!(!task.is_affected(&FxHashSet::default()).unwrap());
+        assert!(!task
+            .is_affected(&FxHashSet::default(), &PathBuf::new())
+            .unwrap());
 
         env::remove_var("BAZ");
     }
@@ -376,6 +383,7 @@ mod is_affected {
     fn returns_true_if_matches_file() {
         let workspace_root = get_fixtures_path("base");
         let project_root = workspace_root.join("files-and-dirs");
+        let project_source = PathBuf::from("files-and-dirs");
         let mut task = create_task(TaskConfig {
             inputs: Some(string_vec!["file.ts"]),
             ..TaskConfig::default()
@@ -384,27 +392,26 @@ mod is_affected {
         task.input_paths.insert(project_root.join("file.ts"));
 
         let mut set = FxHashSet::default();
-        set.insert(project_root.join("file.ts"));
+        set.insert(project_source.join("file.ts"));
 
-        assert!(task.is_affected(&set).unwrap());
+        assert!(task.is_affected(&set, &workspace_root).unwrap());
     }
 
     #[test]
     fn returns_true_if_matches_glob() {
         let workspace_root = get_fixtures_path("base");
-        let project_root = workspace_root.join("files-and-dirs");
+        let project_source = PathBuf::from("files-and-dirs");
         let mut task = create_task(TaskConfig {
             inputs: Some(string_vec!["file.*"]),
             ..TaskConfig::default()
         });
 
-        task.input_globs
-            .insert(glob::normalize(project_root.join("file.*")).unwrap());
+        task.input_globs.insert("files-and-dirs/file.*".into());
 
         let mut set = FxHashSet::default();
-        set.insert(project_root.join("file.ts"));
+        set.insert(project_source.join("file.ts"));
 
-        assert!(task.is_affected(&set).unwrap());
+        assert!(task.is_affected(&set, &workspace_root).unwrap());
     }
 
     #[test]
@@ -418,9 +425,9 @@ mod is_affected {
         task.input_paths.insert(workspace_root.join("package.json"));
 
         let mut set = FxHashSet::default();
-        set.insert(workspace_root.join("package.json"));
+        set.insert(PathBuf::from("package.json"));
 
-        assert!(task.is_affected(&set).unwrap());
+        assert!(task.is_affected(&set, &workspace_root).unwrap());
     }
 
     #[test]
@@ -435,28 +442,28 @@ mod is_affected {
         task.input_paths.insert(project_root.join("file.ts"));
 
         let mut set = FxHashSet::default();
-        set.insert(workspace_root.join("base/other/outside.ts"));
+        set.insert(PathBuf::from("base/other/outside.ts"));
 
-        assert!(!task.is_affected(&set).unwrap());
+        assert!(!task.is_affected(&set, &workspace_root).unwrap());
     }
 
     #[test]
     fn returns_false_if_no_match() {
         let workspace_root = get_fixtures_path("base");
         let project_root = workspace_root.join("files-and-dirs");
+        let project_source = PathBuf::from("files-and-dirs");
         let mut task = create_task(TaskConfig {
             inputs: Some(string_vec!["file.ts", "src/*"]),
             ..TaskConfig::default()
         });
 
         task.input_paths.insert(project_root.join("file.ts"));
-        task.input_globs
-            .insert(glob::normalize(project_root.join("src/*")).unwrap());
+        task.input_globs.insert("files-and-dirs/src/*".into());
 
         let mut set = FxHashSet::default();
-        set.insert(project_root.join("another.rs"));
+        set.insert(project_source.join("another.rs"));
 
-        assert!(!task.is_affected(&set).unwrap());
+        assert!(!task.is_affected(&set, &workspace_root).unwrap());
     }
 
     #[test]
@@ -465,6 +472,7 @@ mod is_affected {
         sandbox.create_file("files-and-dirs/.env", "");
 
         let project_root = sandbox.path().join("files-and-dirs");
+        let project_source = PathBuf::from("files-and-dirs");
         let mut task = create_task(TaskConfig {
             options: TaskOptionsConfig {
                 env_file: Some(TaskOptionEnvFileConfig::Enabled(true)),
@@ -476,8 +484,8 @@ mod is_affected {
         task.input_paths.insert(project_root.join(".env"));
 
         let mut set = FxHashSet::default();
-        set.insert(project_root.join(".env"));
+        set.insert(project_source.join(".env"));
 
-        assert!(task.is_affected(&set).unwrap());
+        assert!(task.is_affected(&set, sandbox.path()).unwrap());
     }
 }